fix(stripe-subscription): finalized docs
