+++ packages/vendure-plugin-stripe-subscription/src/api/subscription-order-item-calculation.ts
@@ -3,44 +3,62 @@ import {
   Order,
   OrderItemPriceCalculationStrategy,
   PriceCalculationResult,
+  ProductVariant,
   RequestContext,
+  UserInputError,
 } from '@vendure/core';
 import { DefaultOrderItemPriceCalculationStrategy } from '@vendure/core/dist/config/order/default-order-item-price-calculation-strategy';
-import { StripeSubscriptionService } from '../api/stripe-subscription.service';
-import {
-  OrderLineWithSubscriptionFields,
-  VariantWithSubscriptionFields,
-} from './subscription-custom-fields';
+import { CustomOrderLineFields } from '@vendure/core/dist/entity/custom-entity-fields';
+import { StripeSubscriptionService } from './stripe-subscription.service';
 
-let subcriptionService: StripeSubscriptionService | undefined;
+let injector: Injector;
 
 export class SubscriptionOrderItemCalculation
   extends DefaultOrderItemPriceCalculationStrategy
   implements OrderItemPriceCalculationStrategy
 {
-  init(injector: Injector): void | Promise<void> {
-    subcriptionService = injector.get(StripeSubscriptionService);
+  init(_injector: Injector): void | Promise<void> {
+    injector = _injector;
   }
 
   // @ts-ignore - Our strategy takes more arguments, so TS complains that it doesnt match the super.calculateUnitPrice
   async calculateUnitPrice(
     ctx: RequestContext,
-    productVariant: VariantWithSubscriptionFields,
-    orderLineCustomFields: OrderLineWithSubscriptionFields['customFields'],
-    order: Order
+    productVariant: ProductVariant,
+    orderLineCustomFields: CustomOrderLineFields,
+    order: Order,
+    orderLineQuantity: number
   ): Promise<PriceCalculationResult> {
-    if (productVariant.customFields.subscriptionSchedule) {
-      const pricing = await subcriptionService!.getPricingForVariant(ctx, {
-        downpayment: orderLineCustomFields.downpayment,
-        startDate: orderLineCustomFields.startDate,
-        productVariantId: productVariant.id as string,
-      });
-      return {
-        price: pricing.amountDueNow,
-        priceIncludesTax: productVariant.listPriceIncludesTax,
-      };
-    } else {
+    const subcriptionService = injector.get(StripeSubscriptionService);
+    if (!subcriptionService) {
+      throw new Error('Subscription service not initialized');
+    }
+    if (!subcriptionService.strategy.isSubscription(ctx, productVariant)) {
       return super.calculateUnitPrice(ctx, productVariant);
     }
+    const subscription = await subcriptionService.strategy.defineSubscription(
+      ctx,
+      injector,
+      productVariant,
+      order,
+      orderLineCustomFields,
+      orderLineQuantity
+    );
+    if (!Array.isArray(subscription)) {
+      return {
+        priceIncludesTax: subscription.priceIncludesTax,
+        price: subscription.amountDueNow ?? 0,
+      };
+    }
+    if (!subscription.length) {
+      throw Error(
+        `Subscription strategy returned an empty array. Must contain atleast 1 subscription`
+      );
+    }
+    const total = subscription.reduce((acc, sub) => sub.amountDueNow || 0, 0);
+    return {
+      priceIncludesTax: subscription[0].priceIncludesTax,
+      price: total,
+    };
   }
 }
