+++ packages/vendure-plugin-stripe-subscription/src/api/vendure-config/custom-fields.ts
