+++ packages/vendure-plugin-stripe-subscription/src/api/vendure-config/stripe-subscription.handler.ts
@@ -7,9 +7,9 @@ import {
   PaymentMethodHandler,
   SettlePaymentResult,
 } from '@vendure/core';
-import { loggerCtx } from '../constants';
-import { printMoney } from './pricing.helper';
-import { StripeSubscriptionService } from './stripe-subscription.service';
+import { loggerCtx } from '../../constants';
+import { StripeSubscriptionService } from '../stripe-subscription.service';
+import { printMoney } from '../util';
 
 let service: StripeSubscriptionService;
 export const stripeSubscriptionHandler = new PaymentMethodHandler({
@@ -18,32 +18,30 @@ export const stripeSubscriptionHandler = new PaymentMethodHandler({
   description: [
     {
       languageCode: LanguageCode.en,
-      value: 'Use a Stripe Subscription as payment',
+      value: 'Stripe Subscription',
     },
   ],
 
   args: {
     apiKey: {
       type: 'string',
-      label: [{ languageCode: LanguageCode.en, value: 'Stripe API key' }],
+      label: [{ languageCode: LanguageCode.en, value: 'API key' }],
       ui: { component: 'password-form-input' },
     },
     publishableKey: {
       type: 'string',
       required: false,
-      label: [
-        { languageCode: LanguageCode.en, value: 'Stripe publishable key' },
-      ],
+      label: [{ languageCode: LanguageCode.en, value: 'Publishable key' }],
       description: [
         {
           languageCode: LanguageCode.en,
-          value:
-            'You can retrieve this via the "eligiblePaymentMethods.stripeSubscriptionPublishableKey" query in the shop api',
+          value: 'For use in the storefront only.',
         },
       ],
     },
     webhookSecret: {
       type: 'string',
+      required: false,
       label: [
         {
           languageCode: LanguageCode.en,
@@ -54,7 +52,7 @@ export const stripeSubscriptionHandler = new PaymentMethodHandler({
         {
           languageCode: LanguageCode.en,
           value:
-            'Secret to validate incoming webhooks. Get this from your Stripe dashboard',
+            'Secret to validate incoming webhooks. Get this from he created webhooks in your Stripe dashboard',
         },
       ],
       ui: { component: 'password-form-input' },
