+++ packages/vendure-plugin-stripe-subscription/src/index.ts
@@ -1,8 +1,9 @@
-export * from './api-v2/generated/graphql';
-export * from './api-v2/stripe-subscription.service';
-export * from './api-v2/strategy/subscription-strategy';
-export * from './api-v2/strategy/default-subscription-strategy';
-export * from './api-v2/vendure-config/has-stripe-subscription-products-payment-checker';
-export * from './api-v2/vendure-config/stripe-subscription.handler';
+export * from './stripe-subscription.plugin';
+export * from './api/generated/graphql';
+export * from './api/stripe-subscription.service';
+export * from './api/strategy/subscription-strategy';
+export * from './api/strategy/default-subscription-strategy';
+export * from './api/vendure-config/has-stripe-subscription-products-payment-checker';
+export * from './api/vendure-config/stripe-subscription.handler';
 export * from './types';
-export * from './api-v2/stripe.client';
+export * from './api/stripe.client';
