+++ packages/vendure-plugin-stripe-subscription/src/ui/stripe-subscription-shared.module.ts
@@ -5,12 +5,10 @@ import {
   registerHistoryEntryComponent,
   SharedModule,
 } from '@vendure/admin-ui/core';
-import { ScheduleRelationSelectorComponent } from './schedule-relation-selector.component';
 import { HistoryEntryComponent } from './history-entry.component';
 
 @NgModule({
   imports: [SharedModule],
-  declarations: [ScheduleRelationSelectorComponent, HistoryEntryComponent],
   providers: [
     registerHistoryEntryComponent({
       type: 'STRIPE_SUBSCRIPTION_NOTIFICATION',
