+++ packages/vendure-plugin-stripe-subscription/test/stripe-test-checkout.plugin.ts
@@ -1,7 +1,7 @@
 import { PluginCommonModule, VendurePlugin } from '@vendure/core';
 import { Body, Controller, Get, Headers, Res } from '@nestjs/common';
 import { Response } from 'express';
-import { clientSecret } from './dev-server';
+import { intent } from './dev-server';
 
 /**
  * Return the Stripe intent checkout page
@@ -14,6 +14,8 @@ export class CheckoutController {
     @Res() res: Response,
     @Body() body: any
   ): Promise<void> {
+    const confirmMethod =
+      intent.intentType === 'SetupIntent' ? 'confirmSetup' : 'confirmPayment';
     res.send(`
 <head>
   <title>Checkout</title>
@@ -36,7 +38,7 @@ export class CheckoutController {
 // See your keys here: https://dashboard.stripe.com/apikeys
 const stripe = Stripe('${process.env.STRIPE_PUBLISHABLE_KEY}');
 const options = {
-  clientSecret: '${clientSecret}',
+  clientSecret: '${intent.clientSecret}',
   // Fully customizable with appearance API.
   appearance: {/*...*/},
 };
@@ -52,8 +54,8 @@ const form = document.getElementById('payment-form');
 form.addEventListener('submit', async (event) => {
   event.preventDefault();
 
-  // const {error} = await stripe.confirmSetup({
-  const {error} = await stripe.confirmPayment({
+  // confirmMethod needs to be 'confirmPayment' or 'confirmSetup'
+  const {error} = await stripe.${confirmMethod}({
     //\`Elements\` instance that was used to create the Payment Element
     elements,
     confirmParams: {
